refactor: package 이동 (line 관련 테스트 -> domain.line)
